Move executeAndPrint to DiffPrinter

+++ scripts/Utils.java
@@ -129,7 +129,7 @@ public class Utils {
         }
     }
 
-    public static void sleep(int millis) {
+    public static void sleep(long millis) {
         try {
             Thread.sleep(millis);
         }
